+++ src/command/DrawImage.cpp
@@ -39,7 +39,9 @@ void DrawImageProgram::initialize(bool commonProgram)
         uniform mat4 uTransform;
 
         out vec2 vUv;
-        out vec2 vLocalPos;
+        // highp so subtraction against uLinearStart in the fragment shader
+        // survives large logical/world coordinates on GLES mediump defaults.
+        out highp vec2 vLocalPos;
 
         void main()
         {
@@ -63,7 +65,7 @@ void DrawImageProgram::initialize(bool commonProgram)
         + (commonProgram ? "#define WHATSCANVAS_COMMON_IMAGE 1\n" : "")
         + wsc::opengl::clipMaskFragmentUniforms() + R"(
         in vec2 vUv;
-        in vec2 vLocalPos;
+        in highp vec2 vLocalPos;
 
         uniform sampler2D uTexture;
         uniform vec4 uTintColor;
@@ -81,9 +83,9 @@ void DrawImageProgram::initialize(bool commonProgram)
         #endif
         uniform int uGradientType;
         uniform int uGradientTileMode;
-        uniform vec2 uLinearStart;
-        uniform vec2 uLinearEnd;
-        uniform vec2 uRadialCenter;
+        uniform highp vec2 uLinearStart;
+        uniform highp vec2 uLinearEnd;
+        uniform highp vec2 uRadialCenter;
         uniform float uRadialRadius;
         uniform int uGradientStopCount;
         uniform float uGradientStopPositions[8];